deploy_python/requirements.txt
progressbar2>=3.43.1
numpy>=1.16.1
